+++ Frontend/zona44_app/lib/bloc/plato_event.dart
@@ -1,0 +1,8 @@
+import 'package:equatable/equatable.dart';
+
+abstract class PlatoEvent extends Equatable {
+  @override
+  List<Object?> get props => [];
+}
+
+class CargarPlatos extends PlatoEvent {}
